Sync in_cksum.c to the same version ospfd has.  This fixes problems with odd packet lengths, which can happen when using TFTP to load a file with an odd length.  ospfd actually took dvmrpd's version in 2006 to fix the same issue, and both daemons implementations are the same.  For the bootloader we keep the consts from the previous version and replace the fatal with a print and return.
ok deraadt@

+++ sys/lib/libsa/net.h
@@ -1,4 +1,4 @@
-/*	$OpenBSD: net.h,v 1.10 2014/11/19 20:28:56 miod Exp $	*/
+/*	$OpenBSD: net.h,v 1.11 2020/05/18 17:01:02 patrick Exp $	*/
 /*	$NetBSD: net.h,v 1.10 1995/10/20 00:46:30 cgd Exp $	*/
 
 /*
@@ -115,7 +115,7 @@ ssize_t	sendrecv(struct iodesc *,
 
 /* Utilities: */
 const char *ether_sprintf(const u_char *);
-int	in_cksum(const void *, int);
+u_int16_t in_cksum(const void *, size_t);
 const char *inet_ntoa(struct in_addr);
 const char *intoa(u_int32_t);		/* similar to inet_ntoa */
 u_int32_t inet_addr(const char *);
